finished last hw

+++ src/app/http/employ/employ.component.ts
@@ -2,26 +2,55 @@ import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { EmployeeService } from 'src/app/employee-service.service';
-import { catchError,tap} from 'rxjs/operators';
+import { catchError,delay,tap} from 'rxjs/operators';
 import { of } from 'rxjs';
-
+import { state, trigger, style, transition, animate } from '@angular/animations';
 
 @Component({
   selector: 'app-employee',
   templateUrl: './employ.component.html',
-  styleUrls: ['./employ.component.scss']
+  styleUrls: ['./employ.component.scss'],
+  animations:[
+    trigger('delete',[
+      state('start',style({
+        opacity:'0',
+        height: '0px',
+        width: '0px',
+        transform: 'translateY(0px)'
+        
+      })),
+      state('end',style({
+        position:'absolute',
+        height: '100px',
+        width: '500px',
+        borderRadius:'5px',
+        transform: 'translateX(440px)',
+        opacity:'1',
+       
+        
+      }
+        
+      )),
+      transition('start => end', animate(1000))
+    ]),
+   
+  ]
+  
 })
 export class EmployeeComponent implements OnInit {
   employee!: FormGroup;
   name!: string;
   salary! : number;
   age!: number;
+  del!:boolean;
+
+  dePop:any='start';
   info: boolean = false;
   checker: boolean = false;
   constructor(private fb: FormBuilder, private http:HttpClient, private EmpService: EmployeeService) {
     
   }
-
+  
   ngOnInit() {
     this.employee = this.fb.group({
       id: ['', Validators.required],
@@ -57,6 +86,7 @@ export class EmployeeComponent implements OnInit {
     this.info = false;
     this.EmpService.getEmp(this.employee.controls.id.value)
     .pipe(
+      delay(1000),
       tap((value:any)=>{
         return this.employee.patchValue(value);
       }),
@@ -72,13 +102,15 @@ export class EmployeeComponent implements OnInit {
     this.employee.reset();
   }
   updateEmp(){
+   
     this.EmpService.update(this.employee.controls.id.value, this.employee.getRawValue())
     .pipe(
       tap( () => {
-        
+       
         this.employee.reset();
         this.checker = false;
         this.info = false;
+        
       }),
       catchError(err => {
         this.ErrorHandler(err);
@@ -87,15 +119,36 @@ export class EmployeeComponent implements OnInit {
     ).subscribe();
   }
   removeEmp(){
-    this.EmpService.delete(this.employee.controls.id.value)
-    .pipe(
-      catchError(e=> {
-        this.ErrorHandler(e);
-        return of([e]);
-      })
-    )
-    .subscribe();
-    this.employee.reset();
+    this.dePop='end';
+   
+    if(this.del===true){
+      
+      this.EmpService.delete(this.employee.controls.id.value)
+      .pipe(
+        delay(2000),
+        catchError(e=> {
+          this.ErrorHandler(e);
+          return of([e]);
+        })
+      )
+      .subscribe();
+      this.employee.reset();
+      this.del=false;
+    }
+    
+   
+  }//
+  yes(){
+   
+    this.dePop='start';
+    this.del=true;
+    this.removeEmp();
+    this.dePop='start';
+    
+  }
+  no(){
+    this.dePop='start';
+    this.del=false;
   }
   
 }
